Split signature into file and remove type_info.rs
